+++ GymMaster/Models/ClientMCards.cs
@@ -10,13 +10,13 @@ public class ClientMCards
     public DateTime? BoughtOnDate { get; private set; }
     public string Barcode { get; private set; }
     public int CurrentEntries { get; private set; }
-    public float PriceSold { get; private set; }
+    public double PriceSold { get; private set; }
     public DateTime? ValidUntil { get; private set; }
     public DateTime? FirstEntry { get; private set; }
     public int GymId { get; private set; }
 
 
-    public ClientMCards(int id, int clientId, int membershipId, DateTime? boughtOnDate, string barcode, int currentEntries, float priceSold, DateTime? validUntil, DateTime? firstEntry, int gymId)
+    public ClientMCards(int id, int clientId, int membershipId, DateTime? boughtOnDate, string barcode, int currentEntries, double priceSold, DateTime? validUntil, DateTime? firstEntry, int gymId)
     {
         Id = id;
         ClientId = clientId;
